backend/manage.py
#!/usr/bin/env python
"""Django's command-line utility for administrative tasks."""
import os
import sys
from dotenv import load_dotenv

from api.connections import GPTConnection, RedisConnection

def main():
    """Run administrative tasks."""

    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'base.settings')
    try:
        from django.core.management import execute_from_command_line
    except ImportError as exc:
        raise ImportError(
            "Couldn't import Django. Are you sure it's installed and "
            "available on your PYTHONPATH environment variable? Did you "
            "forget to activate a virtual environment?"
        ) from exc
    execute_from_command_line(sys.argv)

if __name__ == '__main__':
    dotenv_path = os.path.join(os.path.dirname(__file__), '..', '.env')
    if os.path.exists(dotenv_path):
        load_dotenv()

    print("Creating connection to OpenAI API")
    GPTConnection(
        api_key=os.getenv("OPENAI_API_KEY"),
        url=os.getenv("OPENAI_API_URL")
    )
    print("Connection to OpenAI API created")

    print('Creating connection to Redis')
    RedisConnection(
        host=os.getenv("REDIS_HOST"),
        port=os.getenv("REDIS_PORT")
    )
    print("Connection to Redis created")
    
    main()
