backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.2
tzdata==2024.2
openai==1.65.4
drf-spectacular==0.28.0
pydantic==2.10.3
redis==5.2.1